-misleading prompt during collectstatic script
The collectstatic command runs inside the web container, which prints a warning
using its filesystem (specifically, a path under /srv). The warning doesn't
match the file layout of the working directory, so the message is misleading.
The point of the script is to run the command as scripted, so it's redundant,
too.